--- .github/workflows/run.yml
@@ -1,48 +1,0 @@
-name: 刷步数
-
-on:
-  #push:
-  #  branches: [ main ]
-  #pull_request:
-  #  branches: [ main ]
-  schedule:
-    - cron: '5 0,2,5,7,9,11,13 * * *'
-  watch:
-    types: started
-  workflow_dispatch:
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    if: github.event.repository.owner.id == github.event.sender.id
-
-    steps:
-      - name: Checkout codes
-        uses: actions/checkout@v2
-      - name: Update system and install zsh
-        run: |
-          sudo -E apt-get -qq update
-          sudo -E apt-get install zsh -y
-          
-      - name: 初始化Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.8
-          
-      - name: 开始
-        run: |
-          pip3 install requests
-          user='${{ secrets.USER }}'
-          passwd='${{ secrets.PWD }}'
-          skey='${{ secrets.SKEY }}'
-          sckey='${{ secrets.SCKEY }}'
-          position='${{ secrets.POSITION }}'
-          corpid='${{ secrets.CORPID }}'
-          corpsecret='${{ secrets.CORPSECRET }}'
-          agentid='${{ secrets.AGENTID }}'
-          touser='${{ secrets.TOUSER }}'
-          toparty='${{ secrets.TOPARTY }}'
-          totag='${{ secrets.TOTAG }}'
-          open_get_weather='${{ secrets.OPEN_GET_WEATHER }}'
-          area='${{ secrets.AREA }}'
-          python3 main.py ${user} ${passwd} ${skey} ${sckey} ${position} ${corpid} ${corpsecret} ${agentid} ${touser} ${toparty} ${totag} ${open_get_weather} ${area}
