+++ .gitignore
@@ -1,0 +1,4 @@
+venv
+.vscode
+
+.env
